python check2 removal
